Completed 04_removeFromArray exercise
I really hated this one. I did it in a really really long way. It worked
but was way too long. I also re-wrote and AI-chatted about why and how
the two solution code works. For me, the process I went with was
reductive - like woodworking, you start with something and whittle down
until you get what you want. But the process the solutions took was
additive; start blank and add what you do want to be there at the end.
I'm going to need to have both of these views in the future.

I can't imagine this comment will be helpful for anyone.